GRIDEDIT-247 Changes to get windows build working

+++ libs/MeshKernel/include/MeshKernel/SplineAlgorithms.hpp
@@ -46,14 +46,14 @@ namespace meshkernel
         /// @param[in] startIndex The start spline node
         /// @param[in] endIndex The end spline node
         /// @returns coordinatesDerivatives The second order derivative at corner points
-        [[nodiscard]] static std::vector<Point> SecondOrderDerivative(const std::vector<Point>& splines, UInt startIndex, UInt endIndex);
+        [[nodiscard]] static std::vector<Point> SecondOrderDerivative(const std::vector<Point>& splines, size_t startIndex, size_t endIndex);
 
         /// @brief Second order derivative at spline corner point coordinates (splint)
         /// @param[in] coordinates The spline corner point coordinate (x or y)
         /// @param[in] startIndex The start spline node
         /// @param[in] endIndex The end spline node
         /// @returns coordinatesDerivatives The second order derivative at corner points (x derivative or y derivative)
-        [[nodiscard]] static std::vector<double> SecondOrderDerivative(const std::vector<double>& coordinates, UInt startIndex, UInt endIndex);
+        [[nodiscard]] static std::vector<double> SecondOrderDerivative(const std::vector<double>& coordinates, size_t startIndex, size_t endIndex);
 
         /// @brief Computes curvature in a spline point (comp_curv)
         /// @param[in] splinePoints the spline points
